requirements.txt
setuptools~=56.0.0
requests~=2.25.1
termcolor~=1.1.0
Pillow~=8.2.0